requirements/requirements-testing.txt
factory-boy==2.10.0
pytest==3.4.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-mock==1.7.0
tox==2.9.1